Fix re-authentication when connection bound is on
When re-using a context on a connection, a re-authentication request
may end up trying to use an established context handler to establish
a new context. This will fail with an error in GSSAPI.
Make sure to completely clean up the connection data when a brand
new authentication needs to happen so that no data is mistakenly
carried over.

Note this may leak a small amount of data, but only if authentication is
successful, so it is probably fine as is.

Closes #38

Signed-off-by: Simo Sorce <simo@redhat.com>