Improve mag_conn memory handling
Crate a pool just for the mag_conn structure, so that we can clear up
all the memory used when a reset is necessary.
This also fixes a segfault introduced by a previous patch where we mistakenly
zeroed the whole structure including the memory pool pointer, which needs to
b preserved.

Signed-off-by: Simo Sorce <simo@redhat.com>