Better handling of desired_mechs
If no explicit allowed mechanism is set in configuration just ask
gssapi for a list of known mechanisms and use that. do not try to
artificially acquire credentials as ultimatily all that does is
just call gss_inidicate_mechs() internally.

Do not store the result of gss_inidicate_mechs() on cfg->allowed_mechs
as that would lead to a leak fiven cfg->allowed_mechs is considered
allocated on a memory pool, while gss_inidate_mechs()s results are not.

Signed-off-by: Simo Sorce <simo@redhat.com>