Allow make distcheck to complete successfully
Add an override path option for distcheck so that the shared object can
be installed in a user controlled directory instead of the system
defined apxs one.

make distcheck marks the source directory as read only, this requires
passing some parameters to the tests so that they create the tests
directory in a writable path.
Also add rule to remove these files once done.

Signed-off-by: Simo Sorce <simo@redhat.com>