report file operation errors as warnings
The failures to change permissions and chown the ccache do not prevent
us from continuing to use this ccache. Log messages are confusing to
mod_auth_gssapi users as they leave an impression things are completely
broken.

Change log level to warning to avoid filling up logs with them on error
log level.

Signed-off-by: Alexander Bokovoy <ab@samba.org>
Reviewed-by: Simo Sorce <simo@redhat.com>