Allow sourcing the acceptor_name from httpd
The hostname the client uses to connect to apache is stored in the
request structure for modules to use.
This patch adds an option to specify that the acceptor name to be used
should be constructed on a per-request basis from this name.
This is simply achieved by specifying the literal string {HOSTNAME} to
the existing GssapiAcceptorName option insted of specifying a full
GSSAPI name.

Signed-off-by: Simo Sorce <simo@redhat.com>
Resolves #138