Switch apache mutex type in tests
When running apache as root, it will try to drop privileges.  If the
environment is not perfectly configured, this will result in an
inability to read its own multiprocessing mutex:

    (22)Invalid argument: AH00024: Couldn't set permissions on the
    mpm-accept mutex; check User and Group directives

Since apache also refuses to run as root (-DBIG_SECURITY_HOLE), move
the mutex somewhere that apache will definitely be able to read.

Signed-off-by: Robbie Harwood <rharwood@redhat.com>